Move ternary impls into kernels::ternary
