+++ candle-metal-kernels/src/kernels/mod.rs
@@ -7,6 +7,7 @@ pub mod quantized;
 pub mod reduce;
 pub mod sdpa;
 pub mod sort;
+pub mod ternary;
 pub mod unary;
 
 pub use affine::*;
@@ -17,4 +18,5 @@ pub use quantized::{call_quantized_matmul_mm_t, call_quantized_matmul_mv_t, Ggml
 pub use reduce::*;
 pub use sdpa::{call_sdpa_full, call_sdpa_vector, call_sdpa_vector_2pass};
 pub use sort::{call_arg_sort, call_mlx_arg_sort};
+pub use ternary::call_where_cond_strided;
 pub use unary::*;
