+++ candle-metal-kernels/src/lib.rs
@@ -11,7 +11,8 @@ pub use kernels::{
     affine::*, call_arg_sort, call_binary_contiguous, call_binary_strided, call_cast_contiguous,
     call_cast_strided, call_const_set_contiguous, call_const_set_contiguous_tiled,
     call_const_set_strided, call_mlx_arg_sort, call_mlx_gemm, call_quantized_matmul_mm_t,
-    call_quantized_matmul_mv_t, reduce::*, unary, unary::*, GemmDType, GgmlDType,
+    call_quantized_matmul_mv_t, reduce::*, ternary::call_where_cond_strided, unary, unary::*,
+    GemmDType, GgmlDType,
 };
 use metal::{
     BlitCommandEncoder, Buffer, CommandQueue, ComputeCommandEncoder, ComputePipeline,
@@ -43,56 +44,6 @@ impl DType {
             Self::F32 => 4,
         }
     }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn call_where_cond_strided(
-    device: &Device,
-    ep: impl EncoderProvider,
-    kernels: &Kernels,
-    name: &'static str,
-    shape: &[usize],
-    cond: BufferOffset,
-    cond_stride: &[usize],
-    left: BufferOffset,
-    left_stride: &[usize],
-    right: BufferOffset,
-    right_stride: &[usize],
-    output: &Buffer,
-) -> Result<(), MetalKernelError> {
-    let pipeline = kernels.load_pipeline(device, Source::Ternary, name)?;
-
-    let encoder = ep.encoder();
-    let encoder: &ComputeCommandEncoder = encoder.as_ref();
-    encoder.set_compute_pipeline_state(&pipeline);
-
-    let size: usize = shape.iter().product();
-    let rank = shape.len();
-
-    set_params!(
-        encoder,
-        (
-            size,
-            rank,
-            shape,
-            cond_stride,
-            left_stride,
-            right_stride,
-            &cond,
-            &left,
-            &right,
-            output
-        )
-    );
-
-    let (thread_group_count, thread_group_size) = linear_split(&pipeline, size);
-
-    encoder.use_resource(cond.buffer, MTLResourceUsage::Read);
-    encoder.use_resource(left.buffer, MTLResourceUsage::Read);
-    encoder.use_resource(right.buffer, MTLResourceUsage::Read);
-    encoder.use_resource(output, MTLResourceUsage::Write);
-    encoder.dispatch_thread_groups(thread_group_count, thread_group_size);
-    Ok(())
 }
 
 #[allow(clippy::too_many_arguments)]
